+++ Package.swift
@@ -19,8 +19,8 @@ let package: Package = .init(
     targets: [
         .binaryTarget(
             name: "VComponents",
-            url: "https://github.com/VakhoKontridze/VComponents/releases/download/1.4.3/VComponents.xcframework.zip",
-            checksum: "2982e842418fddd6ee43475954c5c67b17479b10d3a29fd3206235305e1e9e07"
+            url: "https://github.com/VakhoKontridze/VComponents/releases/download/1.4.4/VComponents.xcframework.zip",
+            checksum: "67a40839647d12464c4c65357b0233153c33e96ec974d7a5a8852a2903f97211"
         ),
     ]
 )