+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 ## Change Log
 
-#### [1.4.4(12)]
+#### [1.4.4(12)](https://github.com/VakhoKontridze/VComponents/releases/download/1.4.4/VComponents.xcframework.zip) — *2021 10 28*
 
-- `Resize Indicator` in `VHalfModal` has been renamed to `Grabber`
+- `Resize Indicator` has been renamed to `Grabber` in `VHalfModal`
 - Bug fixes and improvements
 
 #### [1.4.3(11)](https://github.com/VakhoKontridze/VComponents/releases/download/1.4.3/VComponents.xcframework.zip) — *2021 10 11*